+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # CHANGELOG
 
-### [0.1.5] - 2026-08-11
+### [0.1.6] - 2026-08-11
 - Added company manegement functionality.
 - Added company user management functionality.
 - Added company user limit in corporate user management.
+
+### [0.1.5] - 2026-08-09
+- Added orders panel to registered clients.
+- Added order creation to registered clients.
+- Added order search in login view.
+- Added order inbox to order management panel.
 
 ### [0.1.4] - 2026-08-05
 - Added loader when fetching data from the database to improve user experience.